+++ packages/core/__tests__/screens/ListCredentials.test.tsx
@@ -1,14 +1,14 @@
 import { AnonCredsCredentialMetadataKey } from '@credo-ts/anoncreds'
 import { CredentialExchangeRecord, CredentialRole, CredentialState } from '@credo-ts/core'
-import { useCredentialByState } from '@credo-ts/react-hooks'
+import { useCredentialByState } from '@bifold/react-hooks'
 import { useNavigation } from '@react-navigation/native'
 import { act, cleanup, fireEvent, render } from '@testing-library/react-native'
 import React from 'react'
 
-import CredentialCard from '../../src/components/misc/CredentialCard'
 import { StoreProvider, defaultState } from '../../src/contexts/store'
 import ListCredentials from '../../src/screens/ListCredentials'
 import { BasicAppContext } from '../helpers/app'
+import CredentialCardGen from '../../src/components/misc/CredentialCardGen'
 
 interface CredentialContextInterface {
   loading: boolean
@@ -111,7 +111,7 @@ describe('CredentialList Screen', () => {
       </BasicAppContext>,
     )
     await act(async () => {
-      const credentialCards = tree.UNSAFE_getAllByType(CredentialCard)
+      const credentialCards = tree.UNSAFE_getAllByType(CredentialCardGen)
 
       expect(credentialCards).toHaveLength(3)
 
@@ -140,7 +140,7 @@ describe('CredentialList Screen', () => {
       </StoreProvider>,
     )
     await act(async () => {
-      const credentialCards = tree.UNSAFE_getAllByType(CredentialCard)
+      const credentialCards = tree.UNSAFE_getAllByType(CredentialCardGen)
 
       expect(credentialCards).toHaveLength(3)
     })
@@ -163,7 +163,7 @@ describe('CredentialList Screen', () => {
       </StoreProvider>,
     )
     await act(async () => {
-      const credentialCards = tree.UNSAFE_getAllByType(CredentialCard)
+      const credentialCards = tree.UNSAFE_getAllByType(CredentialCardGen)
 
       expect(credentialCards).toHaveLength(3)
     })