+++ packages/core/declarations.d.ts
@@ -1,9 +1,8 @@
 declare module '*.svg' {
   import { SvgProps } from 'react-native-svg'
-  const content: React.FC<SvgProps>
+  import type { FC } from 'react'
+  const content: FC<SvgProps>
   export default content
 }
-
-declare module 'react-native-argon2'
 
 declare module '@react-native-community/netinfo/jest/netinfo-mock'