+++ packages/core/src/screens/ProofRequesting.tsx
@@ -2,7 +2,7 @@ import type { StackScreenProps } from '@react-navigation/stack'
 
 import { ProofCustomMetadata, ProofMetadata, linkProofWithTemplate, sendProofRequest } from '@bifold/verifier'
 import { DidExchangeState, ProofState } from '@credo-ts/core'
-import { useAgent, useProofById } from '@credo-ts/react-hooks'
+import { useAgent, useProofById } from '@bifold/react-hooks'
 import { useFocusEffect, useIsFocused } from '@react-navigation/native'
 import React, { useCallback, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'