+++ packages/core/src/utils/helpers.ts
@@ -25,14 +25,13 @@ import {
   OutOfBandRole,
 } from '@credo-ts/core'
 import { BasicMessageRole } from '@credo-ts/core/build/modules/basic-messages/BasicMessageRole'
-import { useConnectionById } from '@credo-ts/react-hooks'
+import { useConnectionById } from '@bifold/react-hooks'
 import { BrandingOverlay, CaptureBaseAttributeType } from '@bifold/oca'
 import { Attribute, CredentialOverlay, Predicate } from '@bifold/oca/build/legacy'
 import { Buffer } from 'buffer'
 import moment from 'moment'
 import { parseUrl } from 'query-string'
 import { ReactNode } from 'react'
-import { TFunction } from 'react-i18next'
 import { DeviceEventEmitter } from 'react-native'
 
 import { EventTypes, domain } from '../constants'
@@ -59,6 +58,7 @@ import { fallbackDefaultCredentialNameValue, defaultCredDefTag } from './cred-de
 import { getEffectiveCredentialName } from './credential'
 import { isOpenIdCredentialOffer, isOpenIdPresentationRequest } from './parsers'
 import { isMediatorInvitation } from './mediatorhelpers'
+import { TFunction } from 'i18next'
 
 export { parsedCredDefNameFromCredential } from './cred-def'