+++ packages/react-hooks/jest.config.ts
@@ -1,0 +1,24 @@
+import type { Config } from '@jest/types'
+
+import packageJson from './package.json'
+
+const config: Config.InitialOptions = {
+  testTimeout: 120000,
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  coveragePathIgnorePatterns: ['/build/', '/node_modules/', '/__tests__/', 'tests'],
+  coverageDirectory: '<rootDir>/coverage/',
+  verbose: true,
+  testMatch: ['**/?(*.)+(spec|test).[tj]s?(x)'],
+  transform: {
+    '^.+\\.tsx?$': [
+      'ts-jest',
+      {
+        isolatedModules: true,
+      },
+    ],
+  },
+  displayName: packageJson.name,
+}
+
+export default config